requirements.txt
Django==5.2.9
django-constance==4.3.4
django-cors-headers==4.9.0
git+https://github.com/derneuere/django-chunked-upload@master#egg=django-chunked-upload
django-cryptography-5==2.0.3
django-extensions==4.1
django-filter==25.2
django-bulk-update
django-silk==5.4.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
face-recognition==1.3.0
faiss-cpu==1.12.0
Flask==3.1.2
Flask-Cors==6.0.1
Flask-RESTful==0.3.10
geopy==2.4.1
gunicorn==23.0.0
hdbscan==0.8.40
networkx==3.4.2
nltk==3.9.2
markupsafe==3.0.3
Pillow==11.3.0
ImageHash==4.3.1
psycopg==3.2.10
https://github.com/owncloud/pyocclient/archive/master.zip
pytz==2025.2
tzdata==2025.2
PyExifTool==0.4.9
# pyvips==3.0.0
scikit-learn<1.7.3
seaborn==0.13.2
sentence_transformers==2.7.0
timezonefinder==6.6.3
tqdm==4.67.1
gevent==25.9.1
python-magic==0.4.27
Wand==0.6.13
django-q2==1.9.0
safetensors==0.6.2
py-cpuinfo==9.0.0
psutil==7.1.0

# Dependencies for blip
timm==1.0.19
# Dependencies for Moondream chat handler (llama-cpp-python multimodal)
transformers==4.56.2
# Dependencies for mistral quantized and multimodal models like Moondream
llama-cpp-python==0.3.16
argon2-cffi==23.1.0

# Dependencies for image HEIF/AVIF/SVG/RAW support
pillow-heif==1.1.1
cairosvg==2.8.2
rawpy==0.25.1
# For PAM authentication support
django-pam==2.1.2
